fix mppy error

+++ backend/tests/test_ml.py
@@ -277,15 +277,15 @@ class TestRiskManagement:
 class TestDriftMonitoring:
     def test_drift_detection_no_drift(self) -> None:
         dm = DriftMonitor(p_threshold=0.05, psi_threshold=0.2)
-        ref = {"feature": np.array([1.0, 2.0, 3.0, 4.0, 5.0] * 20)}
-        cur = {"feature": np.array([1.0, 2.0, 3.0, 4.0, 5.0] * 20)}
+        ref = {"feature": np.array([1.0, 2.0, 3.0, 4.0, 5.0] * 20, dtype=np.float64)}
+        cur = {"feature": np.array([1.0, 2.0, 3.0, 4.0, 5.0] * 20, dtype=np.float64)}
         results = dm.detect(ref, cur)
         assert all(not r.is_drift for r in results)
 
     def test_correlation_monitor(self) -> None:
         cm = CorrelationMonitor(threshold=0.8)
-        ref = np.array([[1, 2], [3, 4], [5, 6], [7, 8]])
-        cur = np.array([[1, 2], [3, 4], [5, 6], [7, 8]])
+        ref = np.array([[1, 2], [3, 4], [5, 6], [7, 8]], dtype=np.float64)
+        cur = np.array([[1, 2], [3, 4], [5, 6], [7, 8]], dtype=np.float64)
         result = cm.detect(ref, cur)
         assert result.get("significant_change") is False
